KFSPTS-7140: move cab into cam
