+++ src/main/resources/edu/cornell/kfs/cu-spring-kfs.xml
@@ -21,7 +21,6 @@
 	 
 	<import resource="pdp/cu-spring-pdp.xml" />
 	<import resource="module/cam/cu-spring-cam.xml" />
-	<import resource="module/cab/cu-spring-cab.xml" />
 	<import resource="module/ezra/cu-spring-ezra.xml" />
 	<import resource="module/receiptProcessing/cu-spring-receiptProcessing.xml" />
