--- src/main/resources/edu/cornell/kfs/module/cab/cu-spring-cab.xml
@@ -1,44 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p"
-	xmlns:aop="http://www.springframework.org/schema/aop" xmlns:tx="http://www.springframework.org/schema/tx"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-                           http://www.springframework.org/schema/beans/spring-beans-2.0.xsd
-                           http://www.springframework.org/schema/tx
-                           http://www.springframework.org/schema/tx/spring-tx-2.0.xsd
-                           http://www.springframework.org/schema/aop
-                           http://www.springframework.org/schema/aop/spring-aop-2.0.xsd">
-
-	<bean id="cu-cab" parent="cabModuleConfiguration-parentBean">
-		
-		<property name="packagePrefixes">
-            <list merge="true">
-                <value>edu.cornell.kfs.module.cab</value>
-            </list>
-        </property>
-<!--  		<property name="dataDictionaryPackages">
-			<list>
-                <value>edu/cornell/kfs/module/cab/document/datadictionary</value>
-                <value>edu/cornell/kfs/module/cab/businessobject/datadictionary</value>
-			</list>
-		</property>
-		
-		<property name="databaseRepositoryFilePaths">
-			<list>
-				<value>edu/cornell/kfs/module/cab/cu-ojb-cab.xml</value>
-			</list>
-		</property>-->
-		<property name="scriptConfigurationFilePaths">
-			<list/>
-		</property>
-	</bean>
-
-	<bean id="cu-cabModuleService" parent="cabModuleService-parentBean">
-	 	<property name="moduleConfiguration" ref="cu-cab"/>
-	</bean>
-
-    <bean id="capitalAssetBuilderModuleService"  parent="capitalAssetBuilderModuleService-parentBean" 
-       class="edu.cornell.kfs.module.cab.service.impl.CuCapitalAssetBuilderModuleServiceImpl" />
-       
-    <bean id="batchExtractService"  parent="batchExtractService-parentBean" class="edu.cornell.kfs.module.cab.batch.service.impl.CuBatchExtractServiceImpl" />
-
-</beans>
