+++ src/main/resources/institutional-config.properties
@@ -56,7 +56,6 @@ rice.struts.message.resources=\
   org.kuali.kfs.vnd.vnd-resources,\
   org.kuali.kfs.module.ar.ar-resources,\
   org.kuali.kfs.module.bc.bc-resources,\
-  org.kuali.kfs.module.cab.cab-resources,\
   org.kuali.kfs.module.cam.cam-resources,\
   org.kuali.kfs.module.cg.cg-resources,\
   org.kuali.kfs.module.ec.ec-resources,\
@@ -82,7 +81,6 @@ property.files=\
   classpath:org/kuali/kfs/vnd/vnd-resources.properties,\
   classpath:org/kuali/kfs/module/ar/ar-resources.properties,\
   classpath:org/kuali/kfs/module/bc/bc-resources.properties,\
-  classpath:org/kuali/kfs/module/cab/cab-resources.properties,\
   classpath:org/kuali/kfs/module/cam/cam-resources.properties,\
   classpath:org/kuali/kfs/module/cg/cg-resources.properties,\
   classpath:org/kuali/kfs/module/ec/ec-resources.properties,\
